requirements.txt
lxml==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Scrapy==2.11.2
SQLAlchemy==2.0.32
pandas

